selftests: make recent_mock() available to more tests
This simply moves the recent_mock() function to a common location so
that it can be used by other test modules.
