+++ selftests/unit/test_utils_cpu.py
@@ -1,5 +1,4 @@
 import io
-import sys
 import unittest
 
 try:
@@ -7,27 +6,8 @@ try:
 except ImportError:
     import mock
 
-
+from .. import recent_mock
 from avocado.utils import cpu
-
-
-def recent_mock():
-    '''
-    Checks if a recent and capable enough mock library is available
-
-    On Python 2.7, it requires at least mock version 2.0.  On Python 3,
-    mock from the standard library is used, but Python 3.6 or later is
-    required.
-
-    Also, it assumes that on a future Python major version, functionality
-    won't regress.
-    '''
-    if sys.version_info[0] < 3:
-        major = int(mock.__version__.split('.')[0])
-        return major >= 2
-    elif sys.version_info[0] == 3:
-        return sys.version_info[1] >= 6
-    return True
 
 
 class Cpu(unittest.TestCase):
